refactor: 修复部分 circular deps
