+++ packages/s2-core/src/common/interface/export.ts
@@ -1,6 +1,7 @@
-import type { CellMeta, DataItem } from '..';
-import { EXTRA_FIELD } from '..';
 import type { SpreadSheet } from '../../sheet-type';
+import { EXTRA_FIELD } from '../constant';
+import type { CellMeta } from '../interface/interaction';
+import type { DataItem } from '../interface/s2DataConfig';
 
 export type MatrixPlainTransformer = (
   data: DataItem[][],
